// packages/client/src/models/VersionDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VersionDetail = {
  id: number
  version: number
  label: string | null
  description: string | null
  defaultLocationId: string
  defaultLocationName: string | null
  priceCount: number
  isCurrent: boolean
  isLatest: boolean
  createdAt: string
  promotedAt: string | null
  createdByUserId: number | null
}